+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 * Automatic refresh of data now executed one minute after cron event has executed, to let time to do the actions
+* Mini dashboard: Remove javascript, fix margins, add link to main dashboard
 
 ### 1.4.4: May 27th, 2021
 * Dialog Insight: Make birthdate value optional when sending to web service